requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
google-auth-oauthlib>=1.0.0
